Organize existing SQL-logic by category
