+++ back-end/src/graphql/mutation/RootMutation.ts
@@ -2,8 +2,8 @@ import fs from 'fs';
 import { GraphQLBoolean, GraphQLFloat, GraphQLInt, GraphQLList, GraphQLNonNull, GraphQLObjectType, GraphQLString } from 'graphql';
 import { BASE_FILE_UPLOAD_DIRECTORY, pool } from '../..';
 import { SessionData } from '../../model/SessionData';
-import { removeRecipe } from '../../sql/RemoveRecipe';
-import { updatePurchaseState } from '../../sql/UpdatePurchaseState';
+import { removeRecipe } from '../../sql/recipe/RemoveRecipe';
+import { updatePurchaseState } from '../../sql/menu/UpdatePurchaseState';
 import { writeMenuChangeToDatabase } from './helpers/WriteMenuChangeToDatabase';
 import { ModifyMenuResponse } from './ModifyMenuResponse';
 import { ModifyStorage } from './ModifyStorageResponse';
